Add functionality to edit comment in a PR
With this commit, we can now edit comment on the PR. To accomplish this
we have added new endpoints in ui/fork.py and made changes in
templates/_formhelper.html. This commits also adds a file
pull_request_comment_update.html which is a form for editing comments.
Function for the same functionality is added in pagure/lib/__init__.py.

We also added two new fields in pull_request_comments table in DB:
editor_id, edited_on. Corresponding changes in _formheper.html are included to display 'last edited by' in the UI.